+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - In review mode, it's now always possible to navigate in every modules (sequence/subsequence) through the navigation menu
+### Removed
+
+- Remove documentation and server they were no longer used
 
 ## [3.7.0](https://github.com/InseeFr/Drama-Queen/releases/tag/3.7.0) - 2026-07-08